tokenizer/requirements.txt
accelerate==1.0.1
braceexpand==0.1.7
datasets==3.1.0
diffusers==0.33.1
einops==0.8.1
huggingface_hub==0.29.1
iopath==0.1.10
numpy==1.24.3
omegaconf==2.1.1
opencv_python==4.1.2.30
opencv_python_headless==4.11.0.86
Pillow==10.4.0
Requests==2.32.3
scipy==1.10.1
scikit-image==0.21.0
termcolor==2.4.0
torch==2.1.0+cu118
torch_fidelity==0.3.0
torchinfo==1.8.0
torchvision==0.16.0+cu118
tqdm==4.67.1
transformers==4.37.2
webdataset==0.2.100
timm==1.0.15
gdown==5.2.0
open-clip-torch==2.31.0
wandb==0.20.1
